refactor cria uma pasta para o provider e context
