+++ app/frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import Main from './components/main'
-import Provider from './components/provider'
+import Provider from './components/providers/provider'
 
 const App = () => {
   return (
