+++ app/frontend/src/components/eventos/buttonsUp.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react'
 import PropTypes from 'prop-types'
-import Context from '../context'
+import Context from '../providers/context'
 
 const ButtonsUp = (props) => {
   const {
